requirements.txt
google-cloud-aiplatform>=1.49.0
google-generativeai>=0.3.2